api/requirements.txt
tensorflow-macos
tensorflow-metal
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy